refactor: config 하위 config 파일 제거
