+++ src/test/java/com/project/dugoga/config/DataJpaTestBase.java
@@ -1,4 +1,4 @@
-package com.project.dugoga.config.config;
+package com.project.dugoga.config;
 
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.context.annotation.Import;
