+++ src/test/java/com/project/dugoga/config/generator/AvailableAddressFixtureGenerator.java
@@ -1,4 +1,4 @@
-package com.project.dugoga.config.config.generator;
+package com.project.dugoga.config.generator;
 
 import com.project.dugoga.domain.availableaddress.domain.model.entity.AvailableAddress;
