+++ src/test/java/com/project/dugoga/config/generator/OrderFixtureGenerator.java
@@ -1,4 +1,4 @@
-package com.project.dugoga.config.config.generator;
+package com.project.dugoga.config.generator;
 
 import com.project.dugoga.domain.order.domain.model.entity.Order;
 import com.project.dugoga.domain.order.domain.model.enums.OrderStatus;
