+++ src/test/java/com/project/dugoga/config/generator/ProductFixtureGenerator.java
@@ -1,4 +1,4 @@
-package com.project.dugoga.config.config.generator;
+package com.project.dugoga.config.generator;
 
 import com.project.dugoga.domain.product.domain.model.entity.Product;
 import com.project.dugoga.domain.store.domain.model.entity.Store;
