+++ src/test/java/com/project/dugoga/domain/user/application/service/UserServiceTest.java
@@ -18,7 +18,7 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 
 import java.util.Optional;
 
-import static com.project.dugoga.config.config.generator.UserFixtureGenerator.*;
+import static com.project.dugoga.config.generator.UserFixtureGenerator.*;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.Mockito.when;
